Added support for extensions, replaced Hockey on Fastlane
