+++ Wire-iOS Share Extension/Wire-iOS Share Extension-Bridging-Header.h
@@ -4,4 +4,3 @@
 
 
 #import "NSItemProvider+Helper.h"
-#import "PreprocessorHelper.h"
