+++ Wire-iOS/Wire-iOS-Prefix.pch
@@ -28,7 +28,4 @@
 #define NS_DESIGNATED_INITIALIZER
 #endif
 
-#define STRINGIZE_INTERNAL(x) #x
-#define STRINGIZE(x) STRINGIZE_INTERNAL(x)
-
 #endif
