+++ WireCommonComponents/ConfigurationHelper.m
